nlp-sql-in-a-box_Oracle/requirements.txt
openai
python-dotenv
tiktoken
semantic-kernel==0.2.7.dev0
Jinja2
pyodbc
streamlit